Add in $bart_all_working flag to avoid flapping.
BART's API is now intermittently using {} ("all working") as a transition from
A->B, so now we have A->{}->B.  This results in a lot of spurious notifications,
so I'm using a global flag to avoid that behavior.

Longer term, maybe this'd be better handled by moving to sidekiq or similar, but
this is the faster fix.