Add an alias-elevator hack
Elevators now have an 'alias_id' column; if this column is empty, it's
just an elevator, no special handling.  If it's set, that means the name
of the elevator is just an alias for the elevator whose id is alias_id.

Maintaining these alias_ids, and reconciling outages that have been
opened for aliased elevators, is currently a manual process.